Add match path pattern expression.
