+++ src/codec/test/CMakeLists.txt
@@ -30,6 +30,7 @@ set(CODEC_TEST_LIBS
     $<TARGET_OBJECTS:meta_obj>
     $<TARGET_OBJECTS:process_obj>
     $<TARGET_OBJECTS:expression_obj>
+    $<TARGET_OBJECTS:ast_match_path_obj>
     $<TARGET_OBJECTS:function_manager_obj>
     $<TARGET_OBJECTS:wkt_wkb_io_obj>
     $<TARGET_OBJECTS:agg_function_manager_obj>
