+++ src/common/id/test/CMakeLists.txt
@@ -33,6 +33,7 @@ nebula_add_executable(
       $<TARGET_OBJECTS:fs_obj>
       $<TARGET_OBJECTS:meta_client_stats_obj>
       $<TARGET_OBJECTS:expression_obj>
+      $<TARGET_OBJECTS:ast_match_path_obj>
     LIBRARIES
       follybenchmark 
       boost_regex
@@ -72,6 +73,7 @@ nebula_add_test(
     $<TARGET_OBJECTS:fs_obj>
     $<TARGET_OBJECTS:meta_client_stats_obj>
     $<TARGET_OBJECTS:expression_obj>
+    $<TARGET_OBJECTS:ast_match_path_obj>
   LIBRARIES 
     gtest 
     gtest_main
