+++ src/graph/executor/test/CMakeLists.txt
@@ -5,6 +5,7 @@
 SET(EXEC_QUERY_TEST_OBJS
     $<TARGET_OBJECTS:ws_obj>
     $<TARGET_OBJECTS:expression_obj>
+    $<TARGET_OBJECTS:ast_match_path_obj>
     $<TARGET_OBJECTS:network_obj>
     $<TARGET_OBJECTS:process_obj>
     $<TARGET_OBJECTS:graph_thrift_obj>
@@ -39,6 +40,7 @@ SET(EXEC_QUERY_TEST_OBJS
     $<TARGET_OBJECTS:graph_session_obj>
     $<TARGET_OBJECTS:graph_flags_obj>
     $<TARGET_OBJECTS:parser_obj>
+    $<TARGET_OBJECTS:ast_match_path_obj>
     $<TARGET_OBJECTS:validator_obj>
     $<TARGET_OBJECTS:planner_obj>
     $<TARGET_OBJECTS:plan_obj>
