+++ src/graph/visitor/ExprVisitorImpl.h
@@ -41,6 +41,8 @@ class ExprVisitorImpl : public ExprVisitor {
   void visit(ReduceExpression *expr) override;
   // subscript range expression
   void visit(SubscriptRangeExpression *expr) override;
+  // match path pattern expression
+  void visit(MatchPathPatternExpression *expr) override;
 
  protected:
   using ExprVisitor::visit;
