+++ src/graph/visitor/FoldConstantExprVisitor.h
@@ -82,6 +82,8 @@ class FoldConstantExprVisitor final : public ExprVisitor {
   void visit(ReduceExpression *expr) override;
   // subscript range expression
   void visit(SubscriptRangeExpression *expr) override;
+  // match path pattern expression
+  void visit(MatchPathPatternExpression *expr) override;
 
   void visitBinaryExpr(BinaryExpression *expr);
   Expression *fold(Expression *expr);
