+++ src/graph/visitor/RewriteSymExprVisitor.cpp
@@ -336,5 +336,18 @@ void RewriteSymExprVisitor::visit(SubscriptRangeExpression *expr) {
   }
 }
 
+void RewriteSymExprVisitor::visit(MatchPathPatternExpression *expr) {
+  if (expr->inputProp() != nullptr) {
+    expr->inputProp()->accept(this);
+    if (expr_) {
+      if (expr_->kind() != Expression::Kind::kInputProperty) {
+        hasWrongType_ = true;
+        return;
+      }
+      expr->setInputProp(static_cast<InputPropertyExpression *>(expr_));
+    }
+  }
+}
+
 }  // namespace graph
 }  // namespace nebula
