+++ src/graph/visitor/VidExtractVisitor.h
@@ -103,6 +103,8 @@ class VidExtractVisitor final : public ExprVisitor {
   void visit(ReduceExpression *expr) override;
   // subscript range expression
   void visit(SubscriptRangeExpression *expr) override;
+  // match path pattern expression
+  void visit(MatchPathPatternExpression *expr) override;
 
  private:
   void visitBinaryExpr(BinaryExpression *expr);
