+++ src/graph/visitor/test/CMakeLists.txt
@@ -34,6 +34,7 @@ nebula_add_test(
         $<TARGET_OBJECTS:graph_session_obj>
         ${EXPR_FLAG_DEPS}
         $<TARGET_OBJECTS:parser_obj>
+        $<TARGET_OBJECTS:ast_match_path_obj>
         $<TARGET_OBJECTS:idgenerator_obj>
         $<TARGET_OBJECTS:graph_context_obj>
         $<TARGET_OBJECTS:graph_auth_obj>
