+++ src/kvstore/test/CMakeLists.txt
@@ -26,6 +26,7 @@ set(KVSTORE_TEST_LIBS
     $<TARGET_OBJECTS:time_obj>
     $<TARGET_OBJECTS:stats_obj>
     $<TARGET_OBJECTS:expression_obj>
+    $<TARGET_OBJECTS:ast_match_path_obj>
     $<TARGET_OBJECTS:function_manager_obj>
     $<TARGET_OBJECTS:wkt_wkb_io_obj>
     $<TARGET_OBJECTS:agg_function_manager_obj>
